+++ changelog.md
@@ -102,6 +102,8 @@
 - In `std/dom`, `Interval` is now a `ref object`, same as `Timeout`. Definitions of `setTimeout`,
   `clearTimeout`, `setInterval`, `clearInterval` were updated.
 
+- Deprecated using right curly to enclose pragmas; use '.}'" instead.
+
 ## Standard library additions and changes
 
 - `strformat`: